Rename BaseWidgetProvider.kt to WatchWidgetProvider.kt
